generar ventas y reporte de venta

+++ src/ajax.php
@@ -1,10 +1,12 @@
 <?php
 require_once "../conexion.php";
 session_start();
+$anio_actual = date('Y');
 if (isset($_GET['q'])) {
     $datos = array();
     $nombre = $_GET['q'];
-    $cliente = mysqli_query($conexion, "SELECT * FROM cliente WHERE documento LIKE '%$nombre%'");
+   
+    $cliente = mysqli_query($conexion, "SELECT * FROM cliente WHERE documento LIKE '%$nombre%' AND YEAR(fecha) = '$anio_actual'");
     while ($row = mysqli_fetch_assoc($cliente)) {
         $data['id'] = $row['idcliente'];
         $data['label'] = $row['nombre'];
@@ -54,37 +56,55 @@ if (isset($_GET['q'])) {
     die();
 } else if (isset($_GET['procesarVenta'])) {
     $id_cliente = $_GET['id'];
+    $dotaciones = $_GET['dotaciones'];
     $id_user = $_SESSION['idUser'];
-    $consulta = mysqli_query($conexion, "SELECT total, SUM(total) AS total_pagar FROM detalle_temp WHERE id_usuario = $id_user");
-    $result = mysqli_fetch_assoc($consulta);
-    $total = $result['total_pagar'];
-    $insertar = mysqli_query($conexion, "INSERT INTO ventas(id_cliente, total, id_usuario) VALUES ($id_cliente, '$total', $id_user)");
-    if ($insertar) {
-        $id_maximo = mysqli_query($conexion, "SELECT MAX(id) AS total FROM ventas");
-        $resultId = mysqli_fetch_assoc($id_maximo);
-        $ultimoId = $resultId['total'];
-        $consultaDetalle = mysqli_query($conexion, "SELECT * FROM detalle_temp WHERE id_usuario = $id_user");
-        while ($row = mysqli_fetch_assoc($consultaDetalle)) {
-            $id_producto = $row['id_producto'];
-            $cantidad = $row['cantidad'];
-            $desc = $row['descuento'];
-            $precio = $row['precio_venta'];
-            $total = $row['total'];
-            $insertarDet = mysqli_query($conexion, "INSERT INTO detalle_venta (id_producto, id_venta, cantidad, precio, descuento, total) VALUES ($id_producto, $ultimoId, $cantidad, '$precio', '$desc', '$total')");
-            $stockActual = mysqli_query($conexion, "SELECT * FROM producto WHERE codproducto = $id_producto");
-            $stockNuevo = mysqli_fetch_assoc($stockActual);
-            $stockTotal = $stockNuevo['existencia'] - $cantidad;
-            $stock = mysqli_query($conexion, "UPDATE producto SET existencia = $stockTotal WHERE codproducto = $id_producto");
-        } 
-        if ($insertarDet) {
-            $eliminar = mysqli_query($conexion, "DELETE FROM detalle_temp WHERE id_usuario = $id_user");
-            $msg = array('id_cliente' => $id_cliente, 'id_venta' => $ultimoId);
-        } 
-    }else{
-        $msg = array('mensaje' => 'error');
-    }
-    echo json_encode($msg);
-    die();
+    $consultaValidacion = mysqli_query($conexion, "
+       SELECT * FROM ventas 
+        WHERE id_cliente = $id_cliente 
+        AND dataciones = '$dotaciones' 
+        AND YEAR(fecha) = $anio_actual
+    ");
+ 
+    if (mysqli_num_rows($consultaValidacion) > 0) {
+        // Si ya existe, enviar un mensaje de error
+        $msg = array('mensaje' => '2');
+        echo json_encode($msg);
+        die();
+    } else {
+        $consulta = mysqli_query($conexion, "SELECT total, SUM(total) AS total_pagar FROM detalle_temp WHERE id_usuario = $id_user");
+        $result = mysqli_fetch_assoc($consulta);
+        $total = $result['total_pagar'];
+        $insertar = mysqli_query($conexion, "INSERT INTO ventas(id_cliente, total, id_usuario, dataciones) VALUES ($id_cliente, '$total', $id_user, '$dotaciones')");
+    
+        if ($insertar) {
+            $id_maximo = mysqli_query($conexion, "SELECT MAX(id) AS total FROM ventas");
+            $resultId = mysqli_fetch_assoc($id_maximo);
+            $ultimoId = $resultId['total'];
+            $consultaDetalle = mysqli_query($conexion, "SELECT * FROM detalle_temp WHERE id_usuario = $id_user");
+            while ($row = mysqli_fetch_assoc($consultaDetalle)) {
+                $id_producto = $row['id_producto'];
+                $cantidad = $row['cantidad'];
+                $desc = $row['descuento'];
+                $precio = $row['precio_venta'];
+                $total = $row['total'];
+                $insertarDet = mysqli_query($conexion, "INSERT INTO detalle_venta (id_producto, id_venta, cantidad, precio, descuento, total) VALUES ($id_producto, $ultimoId, $cantidad, '$precio', '$desc', '$total')");
+                $stockActual = mysqli_query($conexion, "SELECT * FROM producto WHERE codproducto = $id_producto");
+                $stockNuevo = mysqli_fetch_assoc($stockActual);
+                $stockTotal = $stockNuevo['existencia'] - $cantidad;
+                $stock = mysqli_query($conexion, "UPDATE producto SET existencia = $stockTotal WHERE codproducto = $id_producto");
+            } 
+            if ($insertarDet) {
+                $eliminar = mysqli_query($conexion, "DELETE FROM detalle_temp WHERE id_usuario = $id_user");
+                $msg = array('id_cliente' => $id_cliente, 'id_venta' => $ultimoId);
+            } 
+        }else{
+            $msg = array('mensaje' => '1');
+        }
+
+
+        echo json_encode($msg);
+        die();
+   } 
 }else if (isset($_GET['descuento'])) {
     $id = $_GET['id'];
     $desc = $_GET['desc'];
@@ -150,6 +170,7 @@ else if (isset($_GET['editarColegio'])) {
     exit;
 }
 if (isset($_POST['regDetalle'])) {
+    
     $id = $_POST['id'];
     $cant = $_POST['cant'];
     $precio = $_POST['stock'];
@@ -158,16 +179,36 @@ if (isset($_POST['regDetalle'])) {
     $verificar = mysqli_query($conexion, "SELECT * FROM detalle_temp WHERE id_producto = $id AND id_usuario = $id_user");
     $result = mysqli_num_rows($verificar);
     $datos = mysqli_fetch_assoc($verificar);
+
     if ($result > 0) {
+       
         $cantidad = $datos['cantidad'] + $cant;
         $total_precio = ($cantidad * $total);
         $query = mysqli_query($conexion, "UPDATE detalle_temp SET cantidad = $cantidad, total = '$total_precio' WHERE id_producto = $id AND id_usuario = $id_user");
         if ($query) {
-            $msg = "actualizado";
+                    // Consultar la existencia en la tabla `producto`
+            $consulta_existencia = mysqli_query($conexion, "SELECT existencia FROM producto WHERE codproducto = $id");
+            $producto = mysqli_fetch_assoc($consulta_existencia);
+
+            if ($producto) {
+                $existencia = $producto['existencia'];
+
+                // Verificar si la existencia es menor que la cantidad actualizada
+            if ($existencia < $cantidad) {
+                    // Si no hay suficiente existencia
+                    $msg = "errorexistencia";
+                    
+                // Opcional: puedes deshacer la actualización en `detalle_temp` si deseas
+                mysqli_query($conexion, "UPDATE detalle_temp SET cantidad = cantidad - $cant, total = total - ($cant * $total) WHERE id_producto = $id AND id_usuario = $id_user");
+            } else {
+                // Si todo está correcto
+                $msg = "actualizado";
+            }
         } else {
             $msg = "Error al ingresar";
         }
-    }else{
+    }
+  }else{
         $query = mysqli_query($conexion, "INSERT INTO detalle_temp(id_usuario, id_producto, cantidad ,precio_venta, total) VALUES ($id_user, $id, $cant,'$precio', '$total')");
         if ($query) {
             $msg = "registrado";
@@ -201,5 +242,16 @@ if (isset($_POST['regDetalle'])) {
     echo $msg;
     die();
     
+}else if (isset($_GET['borrartemp'])) {
+    $id = $_SESSION['idUser'];
+    $eliminar = mysqli_query($conexion, "DELETE FROM detalle_temp WHERE id_usuario = $id");
+
+    // Verifica si se eliminó algún registro
+    if (mysqli_affected_rows($conexion) > 0) {
+        echo json_encode(['status' => 'success', 'message' => 'Registro eliminado exitosamente']);
+    } else {
+        echo json_encode(['status' => 'error', 'message' => 'No se encontró ningún registro para eliminar']);
+    }
+    exit;
 }
